Purge old version
Without plans to finish the public version, there's no longer a reason to maintain the unfinished code base especially with the extremely raw state of the last commit. Unfortunately, much of the research that would go into this was done as part of various agreements that mean the code cannot be shared.